refactor: intro duce anti-corruption layer
